Changes to isoaliases() (still messy) and pmagic plugin upgrade

+++ functions.sh
@@ -24,31 +24,39 @@ for i in $MCDDIR/plugins/*;do
 done
 #END LINKS#
 cat $TAGS/linklist|while read i;do
-	IM1=$(echo $i|awk '{print $1}')
-	IM2=$(echo $i|awk '{print $2}')
-	if (echo $i|awk '{print $3}'|grep -q '\.iso') || [ -n "$(echo $i|awk '{print $4}')" ];then
-		echo "More than one matching ISO: $IM1, $IM2 (did not make link)"
-	elif [ -e $IM1 ] && [ ! -e $IM2 ];then
-		if ln -s $IM1 $IM2;then
-			ISOBASENAME=$(echo $IM2|sed -e 's/\.iso//g')
-			touch $TAGS/madelinks #This is to make multicd.sh pause for 1 second so the notifications are readable
-			CUTOUT1=$(echo "$i"|awk 'BEGIN {FS = "*"} ; {print $1}') #The parts of the ISO name before the asterisk
-			CUTOUT2=$(echo "$i"|awk '{print $1}'|awk 'BEGIN {FS = "*"} ; {print $2}') #The parts after the asterisk
-			VERSION=$(echo "$IM1"|awk '{sub(/'"$CUTOUT1"'/,"");sub(/'"$CUTOUT2"'/,"");print}') #Cuts out whatever the asterisk represents (which will be the version number)
-			if [ "$VERSION" != "*" ] && [ "$VERSION" != "$IM1" ];then
-				echo $VERSION > $ISOBASENAME.version #The SystemRescueCD plugin does not use this, but I figure it won't do any harm to have an extra file sitting there.
-				echo "Made a link named $IM2 pointing to $IM1 (version $VERSION)"
-			else	
-				echo "Made a link named $IM2 pointing to $IM1"
-				VERSION="*" #Should remain an asterisk in .defaultname file (see below)
+	DEFAULTNAME=$(echo "$i"|awk '{print $NF}')
+	IM2=$(echo "$i"|awk '{LESS=NF-1; print $LESS}') #What should be linked to
+	IM1=$(echo $i|awk '{LESS=NF-1; for (i=1; i<LESS; i++) print $i }') #Prints all except the last 2 fields. $i is NOT surrounded by quotes, so wildcards are expanded.
+	if !( echo $IM1 | grep -q '\*' ) && [ ! -e $IM2 ];then #IM1 exists (i.e. the asterisk got expanded) and IM2 doesn't exist yet
+		COUNTER=0
+		for j in $IM1;do
+			if [ $COUNTER = 0 ];then
+				LINKTO=$IM2
+			else
+				LINKTO=${COUNTER}_${IM2} #More than one (i.e. Ubuntu)
 			fi
-			if [ -n "$(echo $i|awk '{print $3}')" ];then
-				#The third field of the row will be the default name when multicd.sh asks the user to enter a name.
-				#This could also be used by the menu-writing portion of the plugin script if $TAGS/whatever.name is not present.
-				#Underscores are replaced with spaces. Asterisks are replaced with the $VERSION found above.
-				echo $i|awk '{print $3}'|sed -e 's/_/ /g' -e "s/\*/$VERSION/g">$ISOBASENAME.defaultname
+			if [ -e "$j" ] && ln -s $j $LINKTO;then
+				ISOBASENAME=$(echo $LINKTO|sed -e 's/\.iso//g')
+				touch $TAGS/madelinks #This is to make multicd.sh pause for 1 second so the notifications are readable
+				CUTOUT1=$(echo "$i"|awk 'BEGIN {FS = "*"} ; {print $1}') #The parts of the ISO name before the asterisk
+				CUTOUT2=$(echo "$i"|awk '{print $1}'|awk 'BEGIN {FS = "*"} ; {print $2}') #The parts after the asterisk
+				VERSION=$(echo "$j"|awk '{sub(/'"$CUTOUT1"'/,"");sub(/'"$CUTOUT2"'/,"");print}') #Cuts out whatever the asterisk represents (which will be the version number)
+				if [ "$VERSION" != "*" ] && [ "$VERSION" != "$j" ];then
+					echo $VERSION > $ISOBASENAME.version #The SystemRescueCD plugin does not use this, but I figure it won't do any harm to have an extra file sitting there.
+					echo "Made a link named $LINKTO pointing to $j (version $VERSION)"
+				else	
+					echo "Made a link named $LINKTO pointing to $j"
+					VERSION="*" #Should remain an asterisk in .defaultname file (see below)
+				fi
+				if [ "$DEFAULTNAME" != "none" ];then
+					#The third field of the row will be the default name when multicd.sh asks the user to enter a name.
+					#This could also be used by the menu-writing portion of the plugin script if $TAGS/whatever.name is not present.
+					#Underscores are replaced with spaces. Asterisks are replaced with the $VERSION found above.
+					echo $DEFAULTNAME|sed -e 's/_/ /g' -e "s/\*/$VERSION/g">$ISOBASENAME.defaultname
+				fi
+			COUNTER=$(($COUNTER+1))
 			fi
-		fi
+		done
 	fi
 done
 if [ -f $TAGS/madelinks ];then
